refactor(tui): move status/toast helpers into app/status.rs (#3314 step 3)
Pure code motion for the App god-object split:

- StatusToastLevel and StatusToast (+ its impl) move from app.rs to the
  new app/status.rs, re-exported at the app root so
  crate::tui::app::StatusToast / StatusToastLevel paths are unchanged.
- The toast/status/message cluster leaves the central impl App block as
  an extension impl in status.rs: push_status_toast, set_sticky_status,
  clear_sticky_status, acknowledge_sticky_on_composer_activity,
  sync_status_message_to_toasts, prune_expired_status_toasts,
  active_status_toasts, active_status_toast, classify_status_text,
  is_mode_switch_status_message, and the STICKY_ERROR_TTL_MS constant.
- classify_status_text becomes pub(super) (from private) so the existing
  app/tests.rs assertion on App::classify_status_text keeps compiling —
  the minimal visibility the split requires. No behavior change; the
  state fields stay on App in app.rs.
